Move Author related symbols to lib
